+++ build-logic/src/main/kotlin/GradleHelpers.kt
@@ -1,0 +1,3 @@
+val isOnCI = System.getenv()["GITHUB_ACTIONS"] != null
+
+val KUERY_CLIENT_VERSION = "0.4.0" + if (isOnCI) "" else "-SNAPSHOT"
