+++ kuery-client-compiler/functional-test/build.gradle.kts
@@ -1,7 +1,6 @@
 plugins {
-    id("conventions.kotlin")
-    id("conventions.ktlint")
-    id("conventions.detekt")
+    id("conventions.preset.base")
+    // id("dev.hsbrysk.kuery-client") version "0.4.0-SNAPSHOT"
 }
 
 description = "Kuery client's compiler functional test module."
