+++ kuery-client-spring-data-jdbc/build.gradle.kts
@@ -1,7 +1,5 @@
 plugins {
-    id("conventions.kotlin")
-    id("conventions.ktlint")
-    id("conventions.detekt")
+    id("conventions.preset.base")
     id("conventions.maven-publish")
 }
